Safari-609.1.11.1/LayoutTests/editing/deleting/delete-list-items-in-table-cell-7.html

<!DOCTYPE html>
<html>
<body>
    <div id="container"><table border="1"><tr><td><div contenteditable="true"><ol><li>a</li><li id="li">|b</li></ol></div></td></tr></table></div>
    <script src="../../resources/dump-as-markup.js"></script>
    <script>
        Markup.description('When deleting the second ordered list items in a table cell, which entire content is a contenteditable area, its content is merged with the first list item:');

        getSelection().collapse(document.getElementById("li"), 0);
        Markup.dump('container', 'Before');

        document.execCommand("Delete");
        Markup.dump('container', 'After');
    </script>
</body>
</html>
